+++ data/pom.xml
@@ -32,10 +32,6 @@
 			<artifactId>jackson-databind</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>edu.harvard.hms.dbmi.avillach</groupId>
-			<artifactId>pic-sure-resource-api</artifactId>
-		</dependency>
-		<dependency>
 			<groupId>de.siegmar</groupId>
 			<artifactId>fastcsv</artifactId>
 		</dependency>
