+++ pom.xml
@@ -157,7 +157,7 @@
 			</dependency>
 			<dependency>
 				<groupId>edu.harvard.hms.dbmi.avillach.hpds</groupId>
-				<artifactId>client-api</artifactId>
+				<artifactId>pic-sure-hpds-model</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
@@ -182,30 +182,8 @@
 			</dependency>
 			<dependency>
 				<groupId>edu.harvard.hms.dbmi.avillach</groupId>
-				<artifactId>pic-sure-resource-api</artifactId>
-				<version>2.2.0-SNAPSHOT</version>
-				<exclusions>
-					<exclusion>
-						<groupId>org.apache.logging.log4j</groupId>
-						<artifactId>log4j-core</artifactId>
-					</exclusion>
-					<exclusion>
-						<groupId>org.slf4j</groupId>
-						<artifactId>slf4j-log4j12</artifactId>
-					</exclusion>
-					<exclusion>
-						<groupId>org.slf4j</groupId>
-						<artifactId>slf4j-api</artifactId>
-					</exclusion>
-					<exclusion>
-						<groupId>ch.qos.logback</groupId>
-						<artifactId>logback-classic</artifactId>
-					</exclusion>
-					<exclusion>
-						<groupId>org.slf4j</groupId>
-						<artifactId>slf4j-jdk14</artifactId>
-					</exclusion>
-				</exclusions>
+				<artifactId>pic-sure-api-model</artifactId>
+				<version>3.0.0-SNAPSHOT</version>
 			</dependency>
 			<dependency>
 				<groupId>com.google.guava</groupId>
