+++ service/pom.xml
@@ -23,7 +23,7 @@
 		</dependency>
 		<dependency>
 			<groupId>edu.harvard.hms.dbmi.avillach.hpds</groupId>
-			<artifactId>client-api</artifactId>
+			<artifactId>pic-sure-hpds-model</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>edu.harvard.hms.dbmi.avillach.hpds</groupId>
@@ -32,6 +32,10 @@
 		<dependency>
 			<groupId>edu.harvard.hms.dbmi.avillach.hpds</groupId>
 			<artifactId>common</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>edu.harvard.hms.dbmi.avillach</groupId>
+			<artifactId>pic-sure-api-model</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.apache.commons</groupId>
